Internal #861: Refactor Window Shared Expressions
Move the WindowSharedExpressions class to separate files.
