+++ src/execution/operator/aggregate/physical_window.cpp
@@ -17,6 +17,7 @@
 #include "duckdb/execution/expression_executor.hpp"
 #include "duckdb/function/window/window_executor.hpp"
 #include "duckdb/function/window/window_segment_tree.hpp"
+#include "duckdb/function/window/window_shared_expressions.hpp"
 #include "duckdb/main/client_config.hpp"
 #include "duckdb/main/config.hpp"
 #include "duckdb/parallel/base_pipeline_event.hpp"
