+++ src/function/window/CMakeLists.txt
@@ -1,6 +1,11 @@
 add_library_unity(
-  duckdb_func_window OBJECT window_boundaries_state.cpp window_collection.cpp
-  window_executor.cpp window_segment_tree.cpp)
+  duckdb_func_window
+  OBJECT
+  window_boundaries_state.cpp
+  window_collection.cpp
+  window_executor.cpp
+  window_segment_tree.cpp
+  window_shared_expressions.cpp)
 
 set(ALL_OBJECT_FILES
     ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:duckdb_func_window>
